+++ src/graphql/Schema.ts
@@ -24,12 +24,13 @@ const Schema = gql`
     v: String!
     url: String!
     email: String!
-    isPublished: String!
+    type: String!
+    isPublished: Boolean!
     added: String!
     edited: String!
     published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
+    avgTimeStory: Int
+    avgAllTimeStory: Int
     created: String!
     updated: String!
   }
@@ -37,66 +38,12 @@ const Schema = gql`
     count: Int!
     items: [Article]
   }
-  type Edited @cacheControl(maxAge: 1000) {
-    id: Int!
-    v: String!
-    url: String!
-    email: String!
-    isPublished: String!
-    added: String!
-    edited: String!
-    published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
-    created: String!
-    updated: String!
-  }
-  type EditedMany @cacheControl(maxAge: 1000) {
-    count: Int!
-    items: [Edited]
-  }
   type Filter @cacheControl(maxAge: 1000) {
     id: Int!
     filter: String!
     value: String!
     created: String!
     updated: String!
-  }
-  type Evergreen @cacheControl(maxAge: 1000) {
-    id: Int!
-    v: String!
-    url: String!
-    email: String!
-    isPublished: String!
-    added: String!
-    edited: String!
-    published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
-    created: String!
-    updated: String!
-  }
-  type EvergreenMany @cacheControl(maxAge: 1000) {
-    count: Int!
-    items: [Evergreen]
-  }
-  type TopAuthor @cacheControl(maxAge: 1000) {
-    id: Int!
-    v: String!
-    url: String!
-    email: String!
-    isPublished: String!
-    added: String!
-    edited: String!
-    published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
-    created: String!
-    updated: String!
-  }
-  type TopAuthorMany @cacheControl(maxAge: 1000) {
-    count: Int!
-    items: [TopAuthor]
   }
   type Editor @cacheControl(maxAge: 1000) {
     id: Int!
@@ -209,16 +156,26 @@ const Schema = gql`
     password: String
   }
   ## Article
+  input ArticlePaginationParams {
+    current: Int!
+    limit: Int!
+  }
+  input GetManyArticleParams {
+    type: String!
+    isPublished: Boolean
+    pagination: ArticlePaginationParams
+  }
   input PostOneArticleParams {
     v: String!
     url: String!
     email: String!
-    isPublished: String!
+    type: String!
+    isPublished: Boolean!
     added: String!
     edited: String!
     published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
+    avgTimeStory: Int
+    avgAllTimeStory: Int
   }
   input GetOneArticleParams {
     id: Int!
@@ -227,90 +184,13 @@ const Schema = gql`
     v: String
     url: String
     email: String
-    isPublished: String
+    type: String
+    isPublished: Boolean
     added: String
     edited: String
     published: String
-    avgTimeStory: String
-    avgAllTimeStory: String
-  }
-  ## Edited
-  input PostOneEditedParams {
-    v: String!
-    url: String!
-    email: String!
-    isPublished: String!
-    added: String!
-    edited: String!
-    published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
-  }
-  input GetOneEditedParams {
-    id: Int!
-  }
-  input UpdateOneEditedParams {
-    v: String
-    url: String
-    email: String
-    isPublished: String
-    added: String
-    edited: String
-    published: String
-    avgTimeStory: String
-    avgAllTimeStory: String
-  }
-  ## Evergreen
-  input PostOneEvergreenParams {
-    v: String!
-    url: String!
-    email: String!
-    isPublished: String!
-    added: String!
-    edited: String!
-    published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
-  }
-  input GetOneEvergreenParams {
-    id: Int!
-  }
-  input UpdateOneEvergreenParams {
-    v: String
-    url: String
-    email: String
-    isPublished: String
-    added: String
-    edited: String
-    published: String
-    avgTimeStory: String
-    avgAllTimeStory: String
-  }
-  ## TopAuthor
-  input PostOneTopAuthorParams {
-    v: String!
-    url: String!
-    email: String!
-    isPublished: String!
-    added: String!
-    edited: String!
-    published: String!
-    avgTimeStory: String!
-    avgAllTimeStory: String!
-  }
-  input GetOneTopAuthorParams {
-    id: Int!
-  }
-  input UpdateOneTopAuthorParams {
-    v: String
-    url: String
-    email: String
-    isPublished: String
-    added: String
-    edited: String
-    published: String
-    avgTimeStory: String
-    avgAllTimeStory: String
+    avgTimeStory: Int
+    avgAllTimeStory: Int
   }
   ## Editor
   input PostOneEditorParams {
@@ -486,8 +366,7 @@ const Schema = gql`
   }
   type Query {
     # Get many
-    getManyArticle: ArticleMany!
-    getManyEdited: EditedMany!
+    getManyArticle(where: GetManyArticleParams!): ArticleMany!
     getManyTag: [Tag]!
     getManyQueryS: [QueryS]!
     getManyPage: [Page]!
@@ -496,13 +375,10 @@ const Schema = gql`
     getManyClicksPosition: [ClicksPosition]!
     getManyAppearance: [Appearance]!
     getManyFilter: [Filter]!
-    getManyEvergreen: EvergreenMany!
-    getManyTopAuthor: TopAuthorMany!
     getManyEditor: [Editor]!
     getManyExpandable: [Expandable]!
     # Get one
     getOneUser(where: GetOneUserParams!): User
-    getOneEdited(where: GetOneEditedParams!): Edited
     getOneTag(where: GetOneTagParams!): Tag
     getOneQueryS(where: GetOneQuerySParams!): QueryS
     getOnePage(where: GetOnePageParams!): Page
@@ -513,8 +389,6 @@ const Schema = gql`
     getOneArticle(where: GetOneArticleParams!): Article
     getOneFilter(where: GetOneFilterParams!): Filter
     getOneExpandable(where: GetOneExpandableParams!): Expandable
-    getOneEvergreen(where: GetOneEvergreenParams!): Evergreen
-    getOneTopAuthor(where: GetOneTopAuthorParams!): TopAuthor
     getOneEditor(where: GetOneEditorParams!): Editor
   }
   type Mutation {
@@ -524,8 +398,6 @@ const Schema = gql`
     # Post one
     postOneExpandable(data: PostOneExpandableParams!): Expandable
     postOneTag(data: PostOneTagParams!): Tag
-    postOneEdited(data: PostOneEditedParams!): Edited
-    postOneEvergreen(data: PostOneEvergreenParams!): Evergreen
     postOneQueryS(data: PostOneQuerySParams!): QueryS
     postOnePage(data: PostOnePageParams!): Page
     postOneDevice(data: PostOneDeviceParams!): Device
@@ -534,11 +406,9 @@ const Schema = gql`
     postOneAppearance(data: PostOneAppearanceParams!): Appearance
     postOneFilter(data: PostOneFilterParams!): Filter
     postOneArticle(data: PostOneArticleParams!): Article
-    postOneTopAuthor(data: PostOneTopAuthorParams!): TopAuthor
     postOneEditor(data: PostOneEditorParams!): Editor
     # Update one
     updateOneUser(where: GetOneUserParams!, data: UpdateOneUserParams!): User
-    updateOneEdited(where: GetOneEditedParams!, data: UpdateOneEditedParams!): Edited
     updateOneAppearance(
       where: GetOneAppearanceParams!
       data: UpdateOneAppearanceParams!
@@ -549,12 +419,10 @@ const Schema = gql`
       data: UpdateOneClicksPositionParams!
     ): ClicksPosition
     updateOneArticle(where: GetOneArticleParams!, data: UpdateOneArticleParams!): Article
-    updateOneEvergreen(where: GetOneEvergreenParams!, data: UpdateOneEvergreenParams!): Evergreen
     updateOneExpandable(
       where: GetOneExpandableParams!
       data: UpdateOneExpandableParams!
     ): Expandable
-    updateOneTopAuthor(where: GetOneTopAuthorParams!, data: UpdateOneTopAuthorParams!): TopAuthor
     updateOneEditor(where: GetOneEditorParams!, data: UpdateOneEditorParams!): Editor
     updateOneCountry(where: GetOneCountryParams!, data: UpdateOneCountryParams!): Country
     updateOneDevice(where: GetOneDeviceParams!, data: UpdateOneDeviceParams!): Device
@@ -563,7 +431,6 @@ const Schema = gql`
     updateOneTag(where: GetOneTagParams!, data: UpdateOneTagParams!): Tag
     # Delete one
     deleteOneUser(where: GetOneUserParams!): User
-    deleteOneEdited(where: GetOneEditedParams!): Edited
     deleteOneTag(where: GetOneTagParams!): Tag
     deleteOneQueryS(where: GetOneQuerySParams!): QueryS
     deleteOnePage(where: GetOnePageParams!): Page
@@ -574,8 +441,6 @@ const Schema = gql`
     deleteOneFilter(where: GetOneFilterParams!): Filter
     deleteOneArticle(where: GetOneArticleParams!): Article
     deleteOneExpandable(where: GetOneExpandableParams!): Expandable
-    deleteOneEvergreen(where: GetOneEvergreenParams!): Evergreen
-    deleteOneTopAuthor(where: GetOneTopAuthorParams!): TopAuthor
     deleteOneEditor(where: GetOneEditorParams!): Editor
   }
 `;