+++ src/graphql/resolvers/article/getMany.ts
@@ -1,16 +1,38 @@
 import * as T from '../../../types';
-import { PrismaClient } from '@prisma/client';
+import { PrismaClient, Prisma } from '@prisma/client';
 import type * as GraphQL from '../../../types/graphql';
 
 const prisma = new PrismaClient();
 
 /**
  * Get many articles
+ * @params where [GraphQL.GetManyArticleParams]
  * @return [GraphQL.ArticleMany]
  */
-const getManyArticle: T.Resolver<void, GraphQL.ArticleMany> = async () => {
-  const count = await prisma.article.count();
-  const result = await prisma.article.findMany();
+const getManyArticle: T.Resolver<GraphQL.QueryGetManyArticleArgs, GraphQL.ArticleMany> = async (
+  _,
+  params
+) => {
+  const { where } = params;
+  // Change specifiied where
+  const newWhere: Prisma.ArticleWhereInput = {
+    type: where.type,
+    isPublished: typeof where.isPublished === 'boolean' ? where.isPublished : undefined,
+  };
+  // Get count all with where filter
+  const count = await prisma.article.count({ where: newWhere });
+  // Get pagination values
+  let skip = undefined;
+  let take = undefined;
+  if (where.pagination) {
+    take = where.pagination.limit;
+    skip = take * where.pagination.current - take + 1;
+  }
+  const result = await prisma.article.findMany({
+    where: newWhere,
+    take,
+    skip,
+  });
   return {
     count,
     items: result.map((result) => {
@@ -18,13 +40,14 @@ const getManyArticle: T.Resolver<void, GraphQL.ArticleMany> = async () => {
         id: result.id,
         url: result.url,
         v: result.v,
+        type: result.type,
         email: result.email,
         isPublished: result.isPublished,
-        added: result.added,
-        edited: result.edited,
-        published: result.published,
-        avgTimeStory: result.avgAllTimeStory,
-        avgAllTimeStory: result.avgAllTimeStory,
+        added: result.added.toISOString(),
+        edited: result.edited?.toISOString() || '',
+        published: result.published?.toISOString() || '',
+        avgTimeStory: result.avgAllTimeStory || null,
+        avgAllTimeStory: result.avgAllTimeStory || null,
         created: result.created?.toISOString() || '',
         updated: result.updated?.toISOString() || '',
       };