+++ src/graphql/resolvers/evergreen/index.ts
@@ -1,9 +1,0 @@
-/**
- * Export namespace brands
- */
-
-export { default as getMany } from './getMany';
-export { default as getOne } from './getOne';
-export { default as updateOne } from './updateOne';
-export { default as postOne } from './postOne';
-export { default as deleteOne } from './deleteOne';