+++ src/graphql/resolvers/topAuthor/index.ts
@@ -1,5 +1,0 @@
-export { default as postOne } from './postOne';
-export { default as getOne } from './getOne';
-export { default as getMany } from './getMany';
-export { default as updateOne } from './updateOne';
-export { default as deleteOne } from './deleteOne';