+++ src/prisma/migrations/20210302145043_delete_article_clons/migration.sql
@@ -1,0 +1,49 @@
+/*
+  Warnings:
+
+  - You are about to alter the column `added` on the `Article` table. The data in that column could be lost. The data in that column will be cast from `String` to `DateTime`.
+  - You are about to alter the column `edited` on the `Article` table. The data in that column could be lost. The data in that column will be cast from `String` to `DateTime`.
+  - You are about to alter the column `published` on the `Article` table. The data in that column could be lost. The data in that column will be cast from `String` to `DateTime`.
+  - You are about to drop the `Edited` table. If the table is not empty, all the data it contains will be lost.
+  - You are about to drop the `Evergreen` table. If the table is not empty, all the data it contains will be lost.
+  - You are about to drop the `TopAuthor` table. If the table is not empty, all the data it contains will be lost.
+  - Added the required column `type` to the `Article` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- RedefineTables
+PRAGMA foreign_keys=OFF;
+CREATE TABLE "new_Article" (
+    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    "url" TEXT NOT NULL,
+    "v" TEXT NOT NULL,
+    "email" TEXT NOT NULL,
+    "type" TEXT NOT NULL,
+    "isPublished" TEXT NOT NULL,
+    "added" DATETIME NOT NULL,
+    "edited" DATETIME,
+    "published" DATETIME,
+    "avgTimeStory" TEXT NOT NULL,
+    "avgAllTimeStory" TEXT NOT NULL,
+    "created" DATETIME DEFAULT CURRENT_TIMESTAMP,
+    "updated" DATETIME DEFAULT CURRENT_TIMESTAMP
+);
+INSERT INTO "new_Article" ("id", "url", "v", "email", "isPublished", "added", "edited", "published", "avgTimeStory", "avgAllTimeStory", "created", "updated") SELECT "id", "url", "v", "email", "isPublished", "added", "edited", "published", "avgTimeStory", "avgAllTimeStory", "created", "updated" FROM "Article";
+DROP TABLE "Article";
+ALTER TABLE "new_Article" RENAME TO "Article";
+PRAGMA foreign_key_check;
+PRAGMA foreign_keys=ON;
+
+-- DropTable
+PRAGMA foreign_keys=off;
+DROP TABLE "Edited";
+PRAGMA foreign_keys=on;
+
+-- DropTable
+PRAGMA foreign_keys=off;
+DROP TABLE "Evergreen";
+PRAGMA foreign_keys=on;
+
+-- DropTable
+PRAGMA foreign_keys=off;
+DROP TABLE "TopAuthor";
+PRAGMA foreign_keys=on;